% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{ca2ba}
\alias{ca2ba}
\title{estimates basal area from crown area  (all tree spcies)}
\usage{
ca2ba(ca)
}
\arguments{
\item{ca}{}
}
\description{
estimates basal area from crown area  (all tree spcies)
}
\examples{


 ca.<- 10:300
 plot(ca.,ca2ba(ca.),
 xlab="Crown Area (m*m)",
 ylab="Basal Area (m*m)")

}
