% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{ca2ba_broad}
\alias{ca2ba_broad}
\title{estimates basal area from crown area  (broad leaved tree spcies)}
\usage{
ca2ba_broad(ca)
}
\arguments{
\item{ca}{vector of crown area  (broad leaved tree spcies)}
}
\value{
vector of basal area  (broad leaved tree spcies)
}
\description{
estimates basal area from crown area  (broad leaved tree spcies)
}
\examples{
 ca.<- 10:300
 plot(ca.,ca2ba_broad(ca.),
 xlab="Crown Area (m*m)",
 ylab="Basal Area (m*m)",
 main="Broadleaved")

}
