% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{ca2ba_conif}
\alias{ca2ba_conif}
\title{estimates basal area from crown area  (conifer)}
\usage{
ca2ba_conif(ca)
}
\arguments{
\item{ca}{vector of crown area  (conifer)}
}
\value{
vector of basal area  (conifer)
}
\description{
estimates basal area from crown area  (conifer)
}
\examples{
 ca.<- 10:300
 plot(ca.,ca2ba_conif(ca.),
 xlab="Crown Area (m*m)",
 ylab="Basal Area (m*m)",
 main="Conifer")
}
