% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{ca2dbh}
\alias{ca2dbh}
\title{estimates DBH from crown area}
\usage{
ca2dbh(ca)
}
\arguments{
\item{ca}{}
}
\description{
estimates DBH from crown area
}
\examples{
 ca.<- 10:100
 plot(ca.,ca2dbh(ca.),
 xlab="Crown Area (m*m)",
 ylab="Diameter at Breast Height (cm)")



}
