% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{ca2dbh_broad}
\alias{ca2dbh_broad}
\title{estimates DBH from crown area (broadleaved)}
\usage{
ca2dbh_broad(ca)
}
\arguments{
\item{ca}{}
}
\description{
estimates DBH from crown area (broadleaved)
}
\examples{
 ca.<- 10:100
 plot(ca.,ca2dbh(ca.),
 xlab="Crown Area (m*m)",
 ylab="Diameter at Breast Height (cm)",
 main="Broadleaved")
}
