% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{ca2dbh_conif}
\alias{ca2dbh_conif}
\title{estimates DBH from crown area (conifer)}
\usage{
ca2dbh_conif(ca)
}
\arguments{
\item{ca}{}
}
\value{
DBH (conifer)
}
\description{
estimates DBH from crown area (conifer)
}
\examples{
  ca.<- 10:100
 plot(ca.,ca2dbh(ca.),
 xlab="Crown Area (m*m)",
 ylab="Diameter at Breast Height (cm)",
 main="Conifer")

}
