% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuraiyamaK123Rdata.R
\name{datafarame2sf}
\alias{datafarame2sf}
\title{return sf object from data frame including x,y, coordinates}
\usage{
datafarame2sf(t. = k2ttop, xc = "tx", yc = "ty", crs = 6675)
}
\arguments{
\item{t.}{data frame including x,y, coordinates}

\item{xc}{column name of x}

\item{yc}{column name of y}

\item{crs}{crs code}
}
\value{
sf object
}
\description{
return sf object from data frame including x,y, coordinates
}
\examples{
dir.<-"./ttop/" # dir(dir.)
k1ttop<-read.csv(paste0(dir.,"k1_ttop.csv"))
k2ttop<-read.csv(paste0(dir.,"k2_ttop.csv"))
k3ttop<-read.csv(paste0(dir.,"k3_ttop.csv"))

k123ttop<-list(
  k1=datafarame2sf(k1ttop),
  k2=datafarame2sf(k2ttop),
  k3=datafarame2sf(k3ttop)
)
par(mfrow=c(1,3))
sapply(k123ttop,plot)
for (i in 1:3)plot(k123ttop[[i]]["th"],reset=F)
# save(k123ttop,file="k123ttop.RData")
}
