% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k123_allometry.R
\name{DBH_cah}
\alias{DBH_cah}
\title{Estimate DBH (diameter at breast height) from crown area and tree height}
\usage{
DBH_cah(ca, h, a = 31.2147, b = 0.02329, c = 0.42632)
}
\arguments{
\item{ca}{vector of  crown area}

\item{h}{vector of}

\item{a}{a numeric parameter of a ,default "all species"}

\item{b}{b numeric parameter of a  ,default "all species"}

\item{c}{c numeric parameter of a  ,default "all species"}
}
\value{
vector of DBH
}
\description{
"dbh = a * ca ^b * h^c"
  all species   a=31.21470 ; b= 0.02329  ; c=0.42632
  conifer      a=30.53526  ; b= 0.04826 ; c=0.37625
}
\examples{
p<-subset(PP9,vital>0 & crown)
ca<-p$a ; h<- p$v_cah3 ; dbh <- p$dbh
i<-p$conif

par(mfrow=c(1,3))
# all tree species ####
dbh. <-  DBH_cah(ca,h)
plot(dbh,dbh.,xlab="DBH (field)",ylab="DBH=31.21470*ca^0.02329*h^0.42632",main="All species    adj.R^2  0.9524 ")
points(dbh[i],dbh.[i],col="red")
summary(lm(dbh~dbh.-1))
lines(c(0,120),c(0,120))

# Conifer ####
dbh. <-  DBH_cah_conifer(ca[i],h[i])
plot(dbh[i],dbh.,xlab="DBH (field)",ylab="DBH=30.53526*ca^0.04826*h^0.42632",main="Conifer   adj.R^2  0.9583 ",col="red")
summary(lm(dbh[i]~dbh.-1))
lines(c(0,120),c(0,120))

# Bloadleaved ####
dbh. <-  DBH_cah_broadleaved(ca[!i],h[!i])
plot(dbh[!i],dbh.,xlab="DBH (field)",ylab="DBH=10.9143*ca^0.3209*h^0.1056",main="Bloadleaved    adj.R^2  0.9397 ")
summary(lm(dbh[!i]~dbh.-1))
lines(c(0,120),c(0,120))

}
