% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{DBHTreeHeightPlots123}
\alias{DBHTreeHeightPlots123}
\title{figure of ralatinshps between DBH (cm)and Tree Height (m) for k123}
\usage{
DBHTreeHeightPlots123(P)
}
\value{
only plot
}
\description{
figure of ralatinshps between DBH (cm)and Tree Height (m) for k123
}
\examples{
P <- data.frame(k123_field)
DBHTreeHeightPlots123(P)
}
