% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_K123.R
\name{demo_matsunami_k2}
\alias{demo_matsunami_k2}
\title{demonstration of calculation  and plots of k2 (Matsunami 2022)}
\usage{
demo_matsunami_k2(FB)
}
\arguments{
\item{FB}{}
}
\description{
demonstration of calculation  and plots of k2 (Matsunami 2022)
}
\examples{

demo_matsunami_k2(FB)


}
