% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData.R
\docType{data}
\name{df}
\alias{df}
\title{data frame for summary of tree trunk volume estimation model with crown area and tree height}
\format{
A data frame of nls results for all, conifer,broadleaved tree species
slot
\describe{
  \item{n}{number of samples}
  \item{a}{paremeter a}
  \item{b}{paremeter b}
  \item{c}{paremeter c}
  other statistical information ...
  }
}
\source{
from natural forest in Kuaraiayama forest of Gigu university (Sanchikanri Labo)
}
\usage{
df
}
\description{
TrunkVolume=a*(CrownArea^b)*(TreeHeight^c)  (Nakai et al.)
}
\examples{
df
}
\seealso{
\link{TrankVolume_cah_TV},\link{TrankVolume_cah_model}
}
\keyword{datasets}
