% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2.R
\name{dsm_files}
\alias{dsm_files}
\title{dsm_files}
\usage{
dsm_files(
  ext_ = c(5100, 5200, -1400, 1100),
  files = ext_dsm$f,
  extents = ext_dsm[, c("x1", "x2", "y1", "y2")],
  rect_drow = F
)
}
\arguments{
\item{rect_drow}{draw the extent on map}
}
\value{
file names of dsm files including extents from horizontal coordinates
}
\description{
dsm_files
}
\examples{
# you need files "../gis/dsm/extents_dsm.csv"
ext_dsm<-read.csv("../gis/dsm/extents_dsm.csv")
ext_ <- c(5100,5200,-1400,1100)
dsm_files(ext_)
}
