% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R, R/files.R
\name{file_search}
\alias{file_search}
\title{return file paths including a keyword}
\usage{
file_search(ff, keyword)

file_search(ff, keyword)
}
\arguments{
\item{keyword}{}
}
\description{
return file paths including a keyword

return file paths including a keyword
}
\examples{
(ff<-fullpath())
file_search(ff,"k123")
(ff<-fullpath())
file_search(ff,"k123")
}
