% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R, R/files.R
\name{files_keyword}
\alias{files_keyword}
\title{return data frame of no. and file path  with a keyword}
\usage{
files_keyword(f, keyword = "data")

files_keyword(f, keyword = "data")

files_keyword(f, keyword = "data")
}
\arguments{
\item{paths}{vector of file names (full path)}

\item{text}{character of keyword}
}
\value{
data frame of no. and file path  with a keyword in paths

data frame of no. and file path  with a keyword in paths

data frame of no. and file path  with a keyword in paths
}
\description{
return data frame of no. and file path  with a keyword

return data frame of no. and file path  with a keyword

return data frame of no. and file path  with a keyword
}
\examples{

(f<-dir(pattern="\\\\.R$",recursive =TRUE) ) #
keyword <- "ForestTools"    # "library\\(sf"
(f. <- files_keyword (f,keyword ))
i<-11 ;  readLines(f[i]); grep(keyword ,readLines(f[i]))
edit(file.info(f.)$file)

(f<-dir(pattern="\\\\.R$",recursive =TRUE) ) #
keyword <- "library\\\\(sf"
files_keyword (f,keyword )
i<-11 ;  readLines(f[i]); grep(keyword ,readLines(f[i]))


(f<-dir(pattern="\\\\.R$",recursive =TRUE) ) #
keyword <- "ForestTools"    # "library\\(sf"
(f. <- files_keyword (f,keyword ))
i<-11 ;  readLines(f[i]); grep(keyword ,readLines(f[i]))
edit(file.info(f.)$file)
}
