% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R, R/files.R
\name{fullpath}
\alias{fullpath}
\title{All file path in working directory}
\usage{
fullpath(wd = getwd())

fullpath(wd = getwd())
}
\arguments{
\item{wd}{working directory}
}
\value{
data frame of full path, time , size of all files in working directory

data frame of full path, time , size of all files in working directory
}
\description{
All file path in working directory

All file path in working directory
}
\examples{

fs::dir_tree(path =getwd(), recurse = TRUE,type="dir")
fullpath()

fs::dir_tree(path =getwd(), recurse = TRUE,type="dir")
fullpath()
}
