% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R, R/k123_allometry.R
\name{G}
\alias{G}
\title{figure of ralatinshps between DBH (cm)and Tree Height (m) for k123}
\usage{
G()

G()
}
\description{
figure of ralatinshps between DBH (cm)and Tree Height (m) for k123

figure of ralatinshps between DBH (cm)and Tree Height (m) for k123
}
\examples{
G(PP2)
G()
}
