% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{hinokiV}
\alias{hinokiV}
\title{Japan Rinyacho Zaisekishiki Hinoki}
\usage{
hinokiV(DBH, H)
}
\arguments{
\item{DBH}{numeric vector of DBH}

\item{H}{numeric vector of tree heights}
}
\value{
numeric vector of tree trunk volumes
}
\description{
Japan Rinyacho Zaisekishiki Hinoki
}
\examples{

DBH<-10:80
H <- DBH^0.8
hinokiV(DBH,H)
}
