% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2.R
\name{img2ras}
\alias{img2ras}
\title{convert image matrix to raster matrix}
\usage{
img2ras(m)
}
\arguments{
\item{m}{"image" matrix}
}
\value{
"raster" matrix
}
\description{
convert image matrix to raster matrix
}
\examples{

img2ras(img)
}
