% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData.R
\docType{data}
\name{k123_area_vector}
\alias{k123_area_vector}
\title{K123 sites area polygon (sf class)}
\format{
An object of class \code{list} of length 3.
}
\source{
from natural forest in Kuaraiayama forest of Gigu university (Sanchikanri Labo)
}
\usage{
k123_area_vector
}
\description{
list of  simplified sf object from shape file of sites area polygon
}
\details{
#' @format A sf object of  shape file
\describe{
  \item{k1}{Geometry type: POLYGON}
  \item{k2}{Geometry type: POLYGON}
  \item{k3}{Geometry type: POLYGON}
  }
}
\examples{
par(mfrow=c(1,3));sapply(k123_area_vector,plot)
sapply(k123_area_vector,st_area)

}
\keyword{datasets}
