% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData.R
\docType{data}
\name{k123_drone}
\alias{k123_drone}
\title{Forest stand data distinguished with drone orthophoto and DSM}
\format{
A sf object of  shape file (polygon)
data.frame
\describe{
  \item{id}{site number 1,2,3}
  \item{a}{treeID(ForestTools)}
  \item{h}{label number of tree tags}
  \item{sp}{tree height}
  \item{lbl}{crown area}
  \item{dbh}{Trank volume calculated from crown area and tree height}
  \item{plot}{winRadius(ForestTools)}
  \item{geometry}{geometry of tree positions}
 }
}
\source{
From Forest stand (kuraiyamaK123) from natural forest in Kuaraiayama forest of Gigu university
}
\usage{
k123_drone
}
\description{
Drone orthophoto and DSM was generated with Metashape.
}
\examples{
names(k123_drone)
plot(k123_drone["h"])
}
\keyword{datasets}
