% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData.R
\docType{data}
\name{k123_field}
\alias{k123_field}
\title{Forest stand data from field survey}
\format{
A sf object of  shape file (points)
data.frame
\describe{
  \item{plot}{site number 1,2,3}
  \item{label}{tag numbers of each tree}
  \item{id}{id for dorne survey}
  \item{id_ttops}{id for LiDar survey (ForestTools)}
  \item{dbh}{species name}
  \item{id_drone}{id for LiDar survey (ForestTools)}
  \item{vital}{id for LiDar survey (ForestTools)}
  \item{memo}{id for LiDar survey (ForestTools)}
  \item{crown}{id for LiDar survey (ForestTools)}
  \item{conif}{id for LiDar survey (ForestTools)}
  \item{id_drone}{id for LiDar survey (ForestTools)}
  \item{ca}{id for LiDar survey (ForestTools)}
  \item{h}{id for LiDar survey (ForestTools)}
  ...

 }
}
\source{
from natural forest in Kuaraiayama forest of Gigu university (Sanchikanri Labo)
}
\usage{
k123_field
}
\description{
for every trees (dbh>10cm) in each K123 site
}
\details{
""      ""  "vital"    "memo"     "crown"
"conif"    "id_drone" "ca"       "h"        "V_dh"
"V_Dah2"   "V_Dah"    "geometry"
}
\examples{
par(mfrow=c(1,3));sapply(k123_area_vector,plot)
sapply(k123_area_vector,st_area)

names(k123_field)
plot(k123_field["dbh"])
}
\keyword{datasets}
