% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData.R
\docType{data}
\name{k123_lidar}
\alias{k123_lidar}
\title{Forest stand data (ttops) calculated with aerial LiDar (DSM,DTM,DCHM) with ForestTools}
\format{
A sf object of  shape file (points)
data.frame
\describe{
  \item{plot}{site number 1,2,3}
  \item{treeID}{treeID(ForestTools)}
  \item{lbl}{label number of tree tags}
  \item{height}{tree height}
  \item{ca}{crown area}
  \item{v}{Trank volume calculated from crown area and tree height}
  \item{winRadius}{winRadius(ForestTools)}
  \item{geometry}{geometry of tree positions}
 }
}
\source{
LiDar from Gifu prefecture
Forest stand (kuraiyamaK123) from natural forest in Kuaraiayama forest of Gigu university
}
\usage{
k123_lidar
}
\description{
Forest stand data (ttops) calculated with aerial LiDar (DSM,DTM,DCHM) with ForestTools
}
\examples{
names(k123_lidar)
plot(k123_lidar["height"])
}
\keyword{datasets}
