% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuraiyamaK123Rdata.R
\name{k123_RData}
\alias{k123_RData}
\title{return list of  simplified sf object from shape file of every trees points data}
\usage{
k123_RData(d, lbl = "label", sp = "sp", dbh = "dbh", h = "h", v = "vital")
}
\arguments{
\item{d}{simple future (sf) object forest stund}

\item{lbl}{column name of tree labels}

\item{sp}{column name of tree species}

\item{dbh}{column name of diameter at breast height}

\item{h}{column name of tree height}

\item{v}{column name of vital index}
}
\value{
simple future simplified
}
\description{
return list of  simplified sf object from shape file of every trees points data
}
\examples{
#setwd("../K123_2022")

dir.<-"./tree-point"
dir(dir.)
k1<-st_read(dir.,"k1_2022")
k2<-st_read(dir.,"k2_2022")
k3<-st_read(dir.,"k3_2022")
names(k1) ; names(k2) ; names(k3) ;
k123 <- list(k1=k123_RData(k1),k2=k123_RData(k2),k3=k123_RData(k3))
plot(k123$k1)
plot(k123$k2)
plot(k123$k3)
#save(k123,file="K123.RData")
}
