% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuraiyamaK123Rdata.R
\name{k123crown_RData}
\alias{k123crown_RData}
\title{Title}
\usage{
k123crown_RData(
  d = k1crown,
  lbl = "label",
  sp = "sp",
  a = "Area",
  h = "Height"
)
}
\arguments{
\item{v}{}
}
\description{
Title
}
\examples{
 dir.<-"./crown_polygon" #
 dir(dir.)
k1crown <- st_read(dir.,"polygon_k1_2022")
k2crown <- st_read(dir.,"polygon_k2_2022")
k3crown <- st_read(dir.,"polygon_k3_2022")
k123crown <- list(
   k1=k123crown_RData(d=k1crown,lbl="Label1",sp="樹種",a="Area",h="Height"),
   k2=k123crown_RData(k2crown,lbl="label.i",sp="樹種",a="Area",h="Height"),
   k3=k123crown_RData(k3crown,lbl="Label1",sp="樹種",a="Area",h="Height")
   )

for(i in 1:3){
plot(K123crown[[i]])
}
# save(k123crown,file="k123crown.Rdata")
}
