% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuraiyamaK123Rdata.R
\name{k123plot_RData}
\alias{k123plot_RData}
\title{return list of  simplified sf object from shape file of plot outline polygon data}
\usage{
k123plot_RData(k1crown, k2crown, k3crown)
}
\arguments{
\item{k3crown}{}
}
\description{
return list of  simplified sf object from shape file of plot outline polygon data
}
\examples{
dir.<-"./clown_plot_outline_polygon" # dir(dir.)
k1plot<-st_read(dir.,"k1_crown_plot_outlinepolygon")
k2plot<-st_read(dir.,"k2_crown_plot_outlinepolygon")
k3plot<-st_read(dir.,"k3_crown_plot_outlinepolygon")
#k123plot <- list(k1=st_geometry(k1plot),k2=st_geometry(k2plot),k3=st_geometry(k3plot))
par(mfrow=c(1,3));sapply(k123plot,plot)
sapply(k123plot,st_area)
 #save(k123plot,file="K123plot.RData")
}
