% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{kaiki}
\alias{kaiki}
\title{correlation analysis between basal area and crown area}
\usage{
kaiki(
  x = ca[i],
  y = ba[i],
  main = "針葉樹",
  ylab = "胸高断面積(m^2)",
  xlab = "樹冠面積(m^2)",
  Fx = 20,
  Fy = 0.3,
  Rx = 20,
  Ry = 0.25
)
}
\arguments{
\item{x}{numeric vector of crown area}

\item{y}{numeric vector of basal area}

\item{main}{character for title name of plot}

\item{ylab}{character for y label}

\item{xlab}{character for x label}

\item{Fx}{a value of horizontal position of  text "y = a x"}

\item{Fy}{a value of vertical position of  text "y = a x"}

\item{Rx}{a value of horizontal position of  text "Adj. R^2 (P)"}

\item{Ry}{a value of vertical position of  text "Adj. R^2 (P)"}
}
\value{
results of regression
}
\description{
correlation analysis between basal area and crown area
}
\examples{
d <- data.frame(k123_field)
conif <- d$conif==1 ; ca <- d$ca ; ba <- pi*(d$dbh/200)^2
par(mfrow=c(1,3))
#針葉樹
i<-conif
kaiki(x=ca[i],y=ba[i],main="針葉樹",ylab="胸高断面積(m^2)",xlab="樹冠面積(m^2)",
Fx=20,Fy=0.3,Rx=20,Ry=0.25)
#広葉樹
i<-!conif
kaiki(x=ca[i],y=ba[i],main="広葉樹",ylab="胸高断面積(m^2)",xlab="樹冠面積(m^2)",
Fx=40,Fy=0.65,Rx=50,Ry=0.6)
#全体
kaiki(x=ca,y=ba,main="全種",ylab="胸高断面積(m^2)",xlab="樹冠面積(m^2)",
Fx=40,Fy=0.65,Rx=50,Ry=0.6)
}
