% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{lm_vv}
\alias{lm_vv}
\title{statistical information of regression analysis (lm y=ax)}
\usage{
lm_vv(v., v_)
}
\arguments{
\item{v.}{vector of variable}

\item{v_}{vector of variable}
}
\value{
data frame of statistical information
}
\description{
regression of intercept = 0
}
