% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{lm_xy}
\alias{lm_xy}
\title{statistical information of regression analysis (lm y=ax)}
\usage{
lm_xy(x, y)
}
\arguments{
\item{x}{vector of variable}

\item{y}{vector of variable}
}
\value{
data frame of statistical information
}
\description{
(rename from lm_vv)
regression of intercept = 0
}
\examples{
slotNames(TV2)
x<-TV2@conifer_vv$v.
y<-TV2@conifer_vv$v_
plot(x,y)
lm_xy(x,y)
TV2@conifer_reg

}
