% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{mtime_stamp}
\alias{mtime_stamp}
\title{mtime_stamp}
\usage{
mtime_stamp(fn)
}
\arguments{
\item{fn}{vector of file names}
}
\value{
vector of mtime stamp of files
}
\description{
mtime_stamp
}
\examples{
filename <- dir()
data.frame(filename,mtime=mtime_stamp(filename))
}
