% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{nls_model_summary}
\alias{nls_model_summary}
\title{summarize of nls(Nonlinear Least Squares) result}
\usage{
nls_model_summary(nls_res = TV$all)
}
\arguments{
\item{nls_res}{nls(Nonlinear Least Squares) result}
}
\value{
data frame of summarized nls result
}
\description{
summarize of nls(Nonlinear Least Squares) result
}
\examples{
nls_model_summary(TV$all)
nls_model_summary(TV$conifer)
nls_model_summary(TV$broadleaved)
}
