% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{p_value_to_asterisk}
\alias{p_value_to_asterisk}
\title{p_value_to_asterisk}
\usage{
p_value_to_asterisk(p_value)
}
\arguments{
\item{a}{P values}
}
\value{
strings P<0.05 "*", 0.01 "**", 0.001 "***"
}
\description{
p_value_to_asterisk
}
\examples{
p<-c(5.588200e-03,3.021976e-26,1.196543e-97,0.2,0.05,0.01 )
sapply(p,p_value_to_asterisk )
}
