% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_matsunami_k2.R, R/kuraiyamaK123Rdata.R
\name{plot_test}
\alias{plot_test}
\title{plot_test}
\usage{
plot_test(i)

plot_test(i)
}
\description{
plot_test

plot_test
}
\examples{
plot_test(1)
plot_test(2)
plot_test(3)
plot_test(1)
plot_test(2)
plot_test(3)
}
