% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{qgs_path_name}
\alias{qgs_path_name}
\title{Qgis layer name and path}
\usage{
qgs_path_name(filename.qgs)
}
\arguments{
\item{filename.qgs}{}
}
\value{
matrix of
}
\description{
Qgis layer name and path
}
\examples{

filename.qgs <- "./test/qgis/KuraiyamaK123_Qgis_2023.qgs"
edit(qgs_path_name(filename.qgs))
}
