% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuraiyamaK123Rdata.R
\name{ReadShapefile_CrownPolygon}
\alias{ReadShapefile_CrownPolygon}
\title{read shapefile of crown polygon}
\usage{
ReadShapefile_CrownPolygon(filename = "polygon_k1_2022", ring.tag = "id")
}
\arguments{
\item{ring.tag}{}
}
\description{
read shapefile of crown polygon
}
\examples{
 #setwd("../crown_polygon")
fn <- c("polygon_k1_2022","polygon_k2_2022_2","polygon_k1_2022")

require(rgdal)
dir()


}
