% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuraiyamaK123Rdata.R
\name{ReadShapefile_Points}
\alias{ReadShapefile_Points}
\title{read shapefile points}
\usage{
ReadShapefile_Points(filename = "k2_2022_2")
}
\arguments{
\item{filename}{}
}
\value{
data frame of shapefile points
}
\description{
read shapefile points
}
\examples{
k1<- ReadShapefile_Points("k1_2022")
k2<-　ReadShapefile_Points("k2_2022_2")
k3<-  ReadShapefile_Points("k3_2022")
k123_points<-list(k1=k1,k2=k2,K3=k3)
save(k123_points,file="k123_points.RData")
}
