% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{sawaraV}
\alias{sawaraV}
\title{Japan Rinyacho Zaisekishiki Sawara, Hiba}
\usage{
sawaraV(DBH, H)
}
\arguments{
\item{DBH}{numeric vector of DBH}

\item{H}{numeric vector of tree heights}
}
\value{
numeric vector of tree trunk volumes
}
\description{
Japan Rinyacho Zaisekishiki Sawara, Hiba
}
\examples{

DBH<-10:80
H <- DBH^0.8
sawaraV(DBH,H)
}
