% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{sugiV}
\alias{sugiV}
\title{Japan Rinyacho Zaisekishiki Sugi (Cryptomeria japonica)}
\usage{
sugiV(DBH, H)
}
\arguments{
\item{DBH}{numeric vector of DBH}

\item{H}{numeric vector of tree heights}
}
\value{
numeric vector of tree trunk volumes
}
\description{
Japan Rinyacho Zaisekishiki Sugi (Cryptomeria japonica)
}
\examples{

DBH<-10:80
H <- DBH^0.8
sugiV(DBH,H)

}
