% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2.R
\name{surface3dr}
\alias{surface3dr}
\title{surface3dr
draw surface 3d from raster data}
\usage{
surface3dr(r, ...)
}
\arguments{
\item{r}{}
}
\description{
surface3dr
draw surface 3d from raster data
}
\examples{
r<-raster("07ke013_dchm_NA_filled.tif")
plot(r)
open3d()
surface3dr(r,col="green")
}
