% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{TreeHeight_conif}
\alias{TreeHeight_conif}
\title{conifer tree height estimate from DBH (diameter at brest height)}
\usage{
TreeHeight_conif(dbh)
}
\arguments{
\item{dbh}{}
}
\value{
tree height
}
\description{
conifer tree height estimate from DBH (diameter at brest height)
}
\examples{
dbh <- 1:100
h_conif <- TreeHeight_conif(dbh)

plot(dbh,h_conif,type="l",xlab="DBH (cm)",ylab="Tree Height (m)")
}
