% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{TreeHeight_CR}
\alias{TreeHeight_CR}
\title{Relationship between DBH and tree height (Chapman-Richards)}
\usage{
TreeHeight_CR(dbh, a = 22.5, b = 0.046, c = 1.52)
}
\arguments{
\item{dbh}{vector of diameter at breast height (DBH : cm)}

\item{a}{default a=22.5}

\item{b}{default b=0.046}

\item{c}{default c=1.52}
}
\value{
vector of tree height (m)
}
\description{
Tree height-diameter allometry of 75 major tree species in japan.
Kobayashi et al. 2021
}
\examples{
dbh <- 1:100
plot(dbh,TreeHeight_CR(dbh,a=22.5,b=0.046,c=1.52),
xlab="DBH (cm)",ylab="Tree height (m) " ,
main="Relationship between DBH and tree height (Chapman-Richards model)",
sub="Kobayashi et al. 2021, Abies firma ")
P <- data.frame(k123_field)
DBHTreeHeightPlots123(P)
legend(82,15,c("all species","broad leaved","conifer"),
      col=c("blue","black","red"),lty=c(1,2,2),lwd=c(2,2,2))

dbh. <- 1:100
lines(dbh.,TreeHeight_CR(dbh.,a=36.17,b=0.02888,c=0.997),col="blue",lwd=3) # all species
lines(dbh.,TreeHeight_CR(dbh.,a=33.37,b=0.03461,c=1.006751),col="black",lwd=2,lty=2) # broadleaved
lines(dbh.,TreeHeight_CR(dbh.,a =41.26,b =0.02189,c=1.007),col="red",lwd=2,lty=2) # conifere

}
