% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData.R
\docType{data}
\name{TreeHeight_CR_abc}
\alias{TreeHeight_CR_abc}
\title{Parameters of relationship between DBH and tree height (Chapman-Richards)
calculated from natural forest in Kuaraiayama forest of Gigu university}
\format{
An object of class \code{data.frame} with 3 rows and 3 columns.
}
\usage{
TreeHeight_CR_abc
}
\description{
Parameters of relationship between DBH and tree height (Chapman-Richards)
calculated from natural forest in Kuaraiayama forest of Gigu university
}
\examples{
TreeHeight_CR_abc

}
\keyword{data}
