% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{TrunkVolume_bloardleaved}
\alias{TrunkVolume_bloardleaved}
\title{Japan Rinyacho Zaisekishiki bloadleaved species}
\usage{
TrunkVolume_bloardleaved(DBH, H)
}
\arguments{
\item{DBH}{numeric vector of DBH}

\item{H}{numeric vector of tree heights}
}
\value{
numeric vector of tree trunk volumes
}
\description{
Japan Rinyacho Zaisekishiki bloadleaved species
}
\examples{

DBH<-10:80
H <- DBH^0.8
TrunkVolume_bloardleaved(DBH,H)

}
