% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\docType{class}
\name{TrunkVolume_cah-class}
\alias{TrunkVolume_cah-class}
\title{S4 class object for tree trunk volume regression analysis}
\value{
S4 class object of "TrunkVolume_cah"
}
\description{
Vrin: Tree trunk volume calculated from DBH and tree height with Japanese Rinyacho Zaisekishiki (Table of tree trunk volume),
Vcah: Tree trunk volume calculated from crown area and tree height.
The class object "TrunkVolume_cah" is used for regression analysis between Vrin and Vcah.
}
\section{Slots}{

\describe{
\item{\code{conifer_vv}}{data.frame.}

\item{\code{conifer_reg}}{data.frame.}

\item{\code{broadleaved_vv}}{data.frame.}

\item{\code{broadleaved_reg}}{data.frame.}

\item{\code{all_vv}}{data.frame.}

\item{\code{all_reg}}{data.frame.}
}}

\examples{
TVcah <- new("TrunkVolume_cah")
slotNames(TVcah)
TVcah
# save(TVcah, file = "data/TV.RData")
}
