% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{TrunkVolume_cah}
\alias{TrunkVolume_cah}
\title{This calculates tree trun volume from crown area and tree height}
\usage{
TrunkVolume_cah(ca, h, abc = c(4.147e-05, 0.3017, 2.958))
}
\arguments{
\item{ca}{crown area}

\item{h}{tree height}

\item{abc}{parameters default : abc=c(4.139e-05,0.3018000,2.958000) For all tree species in Kuraiyama natural forest.}
}
\value{
trunk volume
}
\description{
Estimation of Coniferous Standing Tree Volume Using Airborne LiDAR and Passive Optical Remote Sensing
(Nakai et al. 2010)
}
\examples{
TrunkVolume_cah(25,30,c(10^-5.19,1.19,3.19))  #Cryptomeria japonica  (Nakai et al. 2010)
#abc_all        <- c(4.139e-05,0.3018000,2.958000) # all tree species in Kuraiyama natural forest.
#abc_conifer    <- c(5.191e-04,0.5569842,2.005132) # conifer
#abc_broadleabed<- c(1.422e-02,0.5073300,0.879260) # broadleaved
 abc_all         <-c(4.147e-05,3.017e-01,2.958e+00)
 abc_conifer     <-c( 0.0003358,0.5014047,2.1866917)
 abc_broadleabed<-c(0.04063, 0.60575,0.38452)
 TrunkVolume_cah(15:25,20:30,abc_conifer)

}
