% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{TrunkVolume_cah_fig}
\alias{TrunkVolume_cah_fig}
\title{Figure of tree trunk volumes (field (Rinya v=f(dbh,h))vs. drone(a ca^b H^c))}
\usage{
TrunkVolume_cah_fig(vv = vv, lm_vv_ = lm_vv_, title_ = "conif")
}
\arguments{
\item{vv}{data.frame(v_=field, v.=drone)}

\item{lm_vv_}{statistical data (lm,n ,a ,b,c,lm_a,lm_b,lm_adj.r.squared,lm_se)}
}
\value{
graphics of relationships two(field vs. drone) tree trunk volumes
}
\description{
This makes figure of relationships between
tree trunk volumes calculated from DBH measured in field survey and
calculated from  crown area  measured with drone ortho ("m^2")"
(old name of this function : ggplot_vv)
}
\examples{
g1<-TrunkVolume_cah_fig(TV2@conifer_vv,TV2@conifer_reg,"Conifer ")
g2<-TrunkVolume_cah_fig(TV2@broadleaved_vv,TV2@broadleaved_reg,"Broadleaved ")
g3<-TrunkVolume_cah_fig(TV2@all_vv,TV2@all_reg,"All tree species")
gridExtra::grid.arrange(g1, g2, g3,nrow=1)
}
