% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{TrunkVolume_cah_model}
\alias{TrunkVolume_cah_model}
\title{Estimate parameters for function of tree trunk volume with crown area and tree height}
\usage{
TrunkVolume_cah_model(d, a = 4.147e-05, b = 0.3017, c = 2.958)
}
\arguments{
\item{d}{data.frame}
}
\value{
TV list of nls results (TV$all, TV$conifer, TV$broadleaved)
}
\description{
TrunkVolume = a * CrownArea^b * TreeHeight^c
}
\examples{
# data set from  k123_field ####
d <- k123_field
sp <- d$sp ; dbh <- d$dbh ;  h <- d$h ;  ca <- d$ca
conif <- d$conif==1 ; crown <- d$crown==1
v <-TrunkVolume_roman(sp,dbh,h)
d <- data.frame(sp,dbh,h,ca,v,conif,crown)
d <- na.omit(d)
head(d)
# all tree species ####

all <- TrunkVolume_cah_model(d)
summary(all)

# coniferous species ####
conifer <- TrunkVolume_cah_model(subset(d,conif))
summary(conifer)

# broadleaved tree species ####
broadleaved <- TrunkVolume_cah_model(subset(d,!conif),a=0.01136,b=0.65086,c=0.78120)
summary(broadleaved)
TV <- list(all=all,conifer=conifer,broadleaved=broadleaved)
TV
# save(TV,file="TrunkVolume_cah_TV.RData")
# load("TrunkVolume_cah_TV.RData")
l <- lapply(TV,nls_model_summary)

df<-data.frame(n=rbind(l$all$n,l$conifer$n,l$broadleaved$n),row.names=c("all","conifer","bloadleaved"))
df<-data.frame(df,rbind(l$all$coef,l$conifer$coef,l$broadleaved$coef))
df<-data.frame(df,rbind(l$all$ast,l$conifer$ast,l$broadleaved$ast),rbind(l$all$P,l$conifer$P,l$broadleaved$P))
df<-data.frame(df,ResidualStandardError=rbind(l$all$se,l$conifer$se,l$broadleaved$se))
df<-data.frame(df,Adjusted_R_squared=rbind(l$all$adjR2,l$conifer$adjR2,l$broadleaved$adjR2))
df
# save(df,file="TrunkVolume_cah_df.RData")
}
