% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{TrunkVolume_hinoki}
\alias{TrunkVolume_hinoki}
\title{Japan Rinyacho Zaisekishiki Hinoki}
\usage{
TrunkVolume_hinoki(DBH, H)
}
\arguments{
\item{DBH}{numeric vector of DBH}

\item{H}{numeric vector of tree heights}
}
\value{
numeric vector of tree trunk volumes
}
\description{
Japan Rinyacho Zaisekishiki Hinoki
}
\examples{

DBH<-10:80
H <- DBH^0.8
TrunkVolume_hinoki(DBH,H)
}
