% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{TrunkVolume_momi}
\alias{TrunkVolume_momi}
\title{Japan Rinyacho Zaisekishiki other species ex. Abies}
\usage{
TrunkVolume_momi(DBH, H)
}
\arguments{
\item{DBH}{numeric vector of DBH}

\item{H}{numeric vector of tree heights}
}
\value{
numeric vector of tree trunk volumes
}
\description{
Japan Rinyacho Zaisekishiki other species ex. Abies
}
\examples{

DBH<-10:80
H <- DBH^0.8
TrunkVolume_momi(DBH,H)

}
