% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{TrunkVolume_roman}
\alias{TrunkVolume_roman}
\title{calculate tree trunk volume from species,dbh, tree height using with function of Japanese governments}
\usage{
TrunkVolume_roman(sp, dbh, h)
}
\arguments{
\item{sp}{character vector of Japanese species name (Katakana)}

\item{dbh}{vector of dbh (cm)}

\item{h}{vector of tree height (m)}
}
\value{
vector of trunk volume
}
\description{
calculate tree trunk volume from species,dbh, tree height using with function of Japanese governments
}
\examples{
TrunkVolume_roman(c("sugi","hinoki","asunaro","buna"),c(50,30,80),c(20,15,25))
}
