% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData.R
\docType{data}
\name{TV}
\alias{TV}
\title{list for tree trunk volume estimation model from crown area and tree height}
\format{
A list of nls results for all, conifer,broadleaved tree species
slot
\describe{
  \item{TV$all}{nls results}
  \item{TV$conifer}{nls results}
  \item{TVbroadleaved}{nls results}
  }
}
\source{
from natural forest in Kuaraiayama forest of Gigu university (Sanchikanri Labo)
}
\usage{
TV
}
\description{
TrunkVolume=a*(CrownArea^b)*(TreeHeight^c)  (Nakai et al.)
}
\examples{
TV$all
TV$conifer
TV$broadleaved
}
\seealso{
\link{TrankVolume_cah_TV2}
}
\keyword{datasets}
