% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData.R
\docType{data}
\name{TV2}
\alias{TV2}
\title{S4 class object for tree trunk volume regression analysis}
\format{
A S4 class object for tree trunk volume regression analysis
slot
\describe{
  \item{@conifer_vv}{data.frame v_: v_rinya, v.; v_cah}
  \item{@conifer_reg}{data.frame}
  \item{@broadleaved_vv}{data.frame}
  \item{@broadleaved_reg}{data.frame}
  \item{@all_vv}{data.frame}
  \item{@all_reg}{data.frame}


  }
}
\source{
from natural forest in Kuaraiayama forest of Gigu university (Sanchikanri Labo)
}
\usage{
TV2
}
\description{
TrunkVolume=a*(CrownArea^b)*(TreeHeight^c)  (Nakai et al.)
}
\examples{
plot(TV2@conifer_vv)
TV2@conifer_reg
}
\seealso{
\link{TrankVolume_cah_TV2}
}
\keyword{datasets}
