% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2.R
\name{xy2n}
\alias{xy2n}
\title{xy2n
cell number from x,y}
\usage{
xy2n(x, y, r)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{r}{raster}
}
\value{
integer, cell number
}
\description{
xy2n
cell number from x,y
}
\examples{
xy2n(x,y,r)
}
