% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2.R
\name{xyz_ext}
\alias{xyz_ext}
\title{crop xyz data matrix or data frame form extent c(x1,x2,y1,y2)}
\usage{
xyz_ext(xyz, ext)
}
\arguments{
\item{ext}{}
}
\value{
xyz data matrix or data frame
}
\description{
crop xyz data matrix or data frame form extent c(x1,x2,y1,y2)
}
\examples{
xyz_ext(dtm)
}
