% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{DBH_cah}
\alias{DBH_cah}
\title{Estimate DBH (diameter at breast height) from crown area and tree height}
\usage{
DBH_cah(ca, h, abc = DBH_cah_abc$all)
}
\arguments{
\item{ca}{vector of  crown area}

\item{h}{vector of}

\item{abc}{a vevtor of parameters abc ,default "all species"}
}
\value{
vector of DBH
}
\description{
"dbh = a * ca ^b * h^c"
 DBH_cah_abc$all,$conifer,$broadleaved
}
\examples{
p<-subset(k123_field,vital>0 & crown==1)
ca<-p$ca ; h<- p$h ; dbh <- p$dbh ; conif <- p$conif==1
DBH_cah_abc

par(mfrow=c(1,3))
i<-conif
# all tree species ####
dbh. <-  DBH_cah(ca,h)
plot(dbh,dbh.,xlab="DBH (field)",ylab="DBH_ca_h",main="All species")
points(dbh[i],dbh.[i],col="red")
(lm.all <- summary(lm(dbh~dbh.-1)))
lines(c(0,120),c(0,120))

# Conifer ####
dbh. <-  DBH_cah(ca[conif],h[conif],abc=DBH_cah_abc$conifer)
plot(dbh[i],dbh.,xlab="DBH (field)",ylab="DBH_ca_h",main="Conifer",col="red")
(lm.conif <- summary(lm(dbh[i]~dbh.-1)))
lines(c(0,120),c(0,120))

# Broadleaved ####
dbh. <-  DBH_cah(ca[!i],h[!i],abc=DBH_cah_abc$broadleaved)
plot(dbh[!i],dbh.,xlab="DBH (field)",ylab="DBH_ca_h",main="Broadleaved")
(lm.broad <-summary(lm(dbh[!i]~dbh.-1)))
lines(c(0,120),c(0,120))
# statistical table ####
adj.r.squared <- c(lm.all$adj.r.squared,lm.broad$adj.r.squared,lm.conif$adj.r.squared)
P <- c(lm.all$coefficients[, "Pr(>|t|)"],lm.broad$coefficients[, "Pr(>|t|)"],lm.conif$coefficients[, "Pr(>|t|)"])
data.frame(t(DBH_cah_abc),adj.r.squared,P)

}
