% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yn.R
\name{TreeHeight}
\alias{TreeHeight}
\title{calculate tree height from species and dbh using with function (Kato)}
\usage{
TreeHeight(conif_sp, sp, dbh)
}
\arguments{
\item{sp}{character vector of Japanese species name (Katakana)}

\item{dbh}{vector of dbh (cm)}
}
\value{
vector of tree height
}
\description{
calculate tree height from species and dbh using with function (Kato)
}
\examples{
conif_sp<-c("アスナロ","スギ","サワラ")
TreeHeight(conif_sp,c("スギ","ヒノキ","ブナ"),c(50,30,80))
}
