% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData.R
\docType{data}
\name{TreeHeight_CR_abc}
\alias{TreeHeight_CR_abc}
\title{Parameters of relationship between DBH and tree height (Chapman-Richards)}
\format{

}
\source{
from natural forest in Kuaraiayama forest of Gigu university (Sanchikanri Labo)
}
\usage{
TreeHeight_CR_abc
}
\description{
H=a*(1-exp(-b*DBH))^c + 1.3 (Chapman-Richards)
}
\seealso{
\link{TreeHeight_CR}
}
\keyword{datasets}
