% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{yield_density}
\alias{yield_density}
\title{Title}
\usage{
yield_density(v, a)
}
\arguments{
\item{v}{volume of trees}

\item{a}{area of plot}
}
\value{
data frame of yn (number of trees ordered with size , yield of cumulative volume)
}
\description{
Title
}
\examples{
pn <- 1
d<-subset(K,plt==pn)
a<-as.numeric(st_area(k123plot[[pn]]))
plot(yield_density(d$v,a),
 xlab="N",
 ylab="Cummulative trunk volume(m^3/ha)")

}
